Adapt validate, cancel and update reservation

+++ app/controllers/reservations_controller.rb
@@ -25,10 +25,8 @@ class ReservationsController < ApplicationController
     @reservation = @advert.reservations.new(reservation_params)
     @reservation.user = current_user
     if @reservation.save
-      body = "Vous avez reçu une demande de réservation de la part de #{@reservation.user.username}\n\nAccédez dès maintenant à votre espace propriétaire pour valider ou annuler cette réservation.\n\nMeilleures salutations,\nToque et Stocke"
-      subject = "Demande de réservation pour l'espace #{@advert.title}"
       recipients = User.where(id: @advert.user_id)
-      current_user.send_message(recipients, body, subject).conversation
+      admin_user.send_message(recipients, @reservation.email_ask[:body], @reservation.email_ask[:subject]).conversation
       flash[:notice] = "Votre demande de réservation a bien été envoyée au propriétaire."
     else
       flash[:alert] = "Les données saisies pour votre réservation sont incorrectes. Veuillez vérifier que les dates choisies sont valides et réessayer."
@@ -40,14 +38,17 @@ class ReservationsController < ApplicationController
   end
   
   def validate
+    return unless current_user == @advert.user
     if @reservation.update_attributes(validated: true)
-      ValidateReservationEmail.perform_async(@reservation.user_id, @reservation.id)
-      FeedbackEmail.perform_at((@reservation.end_date.to_time + 10.hours).to_i, @reservation.advert_id, @reservation.advert.user_id, @reservation.user_id)
+      # ValidateReservationEmail.perform_async(@reservation.user_id, @reservation.id)
+      # FeedbackEmail.perform_at((@reservation.end_date.to_time + 10.hours).to_i, @reservation.advert_id, @reservation.advert.user_id, @reservation.user_id)
+      recipients = User.where(id: @reservation.user_id)
+      admin_user.send_message(recipients, @reservation.email_validate[:body], @reservation.email_validate[:subject]).conversation
       flash[:notice] = "La réservation a bien été validée. Veuillez consulter vos mails pour obtenir un exemplaire du contrat de location."
     else
       flash[:alert] = "Une erreur est survenue durant la validation de la réservation. Veuillez réessayer s'il vous plaît. Si le problème persiste, n'hésitez pas à contacter le support."
     end
-    redirect_to edit_user_registration_path(current_user)
+    redirect_to user_path(current_user)
   end
   
   def preview_cancel
@@ -57,8 +58,11 @@ class ReservationsController < ApplicationController
   end
   
   def update
+    return unless current_user == @reservation.user
     if @reservation.update_attributes(reservation_params)
-      ModifyReservationEmail.perform_async(@reservation.advert.user.id, @reservation.id)
+      # ModifyReservationEmail.perform_async(@reservation.advert.user.id, @reservation.id)
+      recipients = User.where(id: @advert.user_id)
+      admin_user.send_message(recipients, @reservation.email_update[:body], @reservation.email_update[:subject]).conversation
       flash[:notice] = "Votre demande de modification concernant votre réservation a bien été envoyée au propriétaire."
     else
       flash[:alert] = "Une erreur est survenue durant la mise à jour de votre réservation. Veuillez réessayer s'il vous plaît. Si le problème persiste, n'hésitez pas à contacter le support."
@@ -67,8 +71,11 @@ class ReservationsController < ApplicationController
   end
   
   def destroy
+    return unless current_user == @advert.user or current_user == @reservation.user
     if @reservation.destroy
-      CancelReservationEmail.perform_async(@reservation.user_id, @reservation.id)
+      # CancelReservationEmail.perform_async(@reservation.user_id, @reservation.id)
+      recipients = User.where(id: @reservation.user_id)
+      admin_user.send_message(recipients, @reservation.email_cancel[:body], @reservation.email_cancel[:subject]).conversation
       flash[:notice] = "La réservation a bien été annulée. Un email a été envoyé au locataire pour l'en informer."
     else
       flash[:alert] = "Une erreur est survenue durant l'annulation de la réservation. Veuillez réessayer s'il vous plaît. Si le problème persiste, n'hésitez pas à contacter le support."
